Modified Big Button, button_container and title
There is an 8px gap between each of the blue
sections inside the card.
The title of the card uses a 16px font.
There are 8px between the title text and the edge of the title section.
The content section has 16px space on the top and bottom, and 8px on either side
Everything inside the `.button-container` section is centered, and there is 8px padding
The Big Button has 24px space on the sides, and 8px on top and bottom